backend/requirements.txt
uvicorn[standard]
fastapi[standard]
python-multipart
email-validator
passlib[bcrypt]
pydantic
emails
gunicorn
jinja2
alembic
httpx
psycopg[binary]
sqlmodel>=0.0.16
bcrypt
pydantic-settings
sentry-sdk[fastapi]
pyjwt
python-jose
beautifulsoup4
openai
anthropic
sdmx1
tavily-python
sentinelsat
sentinelhub
opol >= 0.1.63
aiohttp
google-generativeai
google-genai>=1.13.0
instructor
pytest
pytest-asyncio
mypy
ruff
fastmcp
pre-commit
types-passlib
minio
bson
python-multipart
pymupdf
celery
redis
eventlet
croniter
sqlalchemy-celery-beat
pytest-celery
watchfiles>=0.18.1
Werkzeug>=3.0.0
pgvector
feedparser
newspaper4k
lxml[html_clean]
uvloop
httptools
openpyxl